Move NameMap to its own file
