+++ flatgfa/src/lib.rs
@@ -4,6 +4,7 @@ pub mod flatgfa;
 pub mod gaf;
 pub mod gfaline;
 pub mod memfile;
+pub mod namemap;
 pub mod parse;
 pub mod pool;
 pub mod print;
