+++ flatgfa/src/parse.rs
@@ -1,8 +1,7 @@
-use crate::flatgfa::{self, Handle, LineKind, Orientation, Segment};
+use crate::flatgfa::{self, Handle, LineKind, Orientation};
 use crate::gfaline;
 use crate::memfile::MemchrSplit;
-use crate::pool::Id;
-use std::collections::HashMap;
+use crate::namemap::NameMap;
 use std::io::BufRead;
 
 pub struct Parser<'a, P: flatgfa::StoreFamily<'a>> {
@@ -176,36 +175,6 @@ impl Parser<'static, flatgfa::HeapFamily> {
 impl<'a> Parser<'a, flatgfa::FixedFamily> {
     pub fn for_slice(store: flatgfa::FixedGFAStore<'a>) -> Self {
         Self::new(store)
-    }
-}
-
-#[derive(Default)]
-pub struct NameMap {
-    /// Names at most this are assigned *sequential* IDs, i.e., the ID is just the name
-    /// minus one.
-    sequential_max: usize,
-
-    /// Non-sequential names go here.
-    others: HashMap<usize, u32>,
-}
-
-impl NameMap {
-    pub fn insert(&mut self, name: usize, id: Id<Segment>) {
-        // Is this the next sequential name? If so, no need to record it in our hash table;
-        // just bump the number of sequential names we've seen.
-        if (name - 1) == self.sequential_max && (name - 1) == id.index() {
-            self.sequential_max += 1;
-        } else {
-            self.others.insert(name, id.into());
-        }
-    }
-
-    pub fn get(&self, name: usize) -> Id<Segment> {
-        if name <= self.sequential_max {
-            ((name - 1) as u32).into()
-        } else {
-            self.others[&name].into()
-        }
     }
 }
